s2-dashboard-template/pyproject.toml
[build-system]
  build-backend = "poetry.core.masonry.api"
  requires = ["poetry-core"]

[project]
  dependencies = ["singlestoredb (>=1.11.0,<2.0.0)","python-dotenv (>=1.0.1,<2.0.0)", "dash (>=2.18.2,<3.0.0)", "psutil (>=7.0.0,<8.0.0)"]
  description = ""
  name = "s2-dashboard-template"
  requires-python = ">=3.9"
  version = "1.0.0"
  [project.scripts]
    local-dev = "s2_dashboard_template:main"
    nova = "s2_dashboard_template:run_app"
